fixing some stuff
